#6 backend works with integers for bowls percentages

+++ core/budgets/entities.py
@@ -5,12 +5,12 @@ from core.utils.numbers import Decimal
 @dataclass
 class Budget:
     amount: Decimal
-    essentials: float
-    education: float
-    goals: float
-    retirement: float
-    loose: float
+    essentials: int
+    education: int
+    goals: int
+    retirement: int
+    loose: int
 
     @property
-    def total_percentage(self) -> float:
+    def total_percentage(self) -> int:
         return self.essentials + self.education + self.goals + self.retirement + self.loose
